refactor: split app layout into components
